1.修改了shader系统结构,移除特殊的内建变量,只留通用全局变量,现在可以支持自己随便写,还能方便的获得全局变量 2.编辑器中的数据不再拥有约束(正在制作约束) 3.代码中可以使用shader类中的方法来获得自定义uniform的index,以便快速更新数据 2.基本的模型加载 3.基本的资源管理器

+++ Engine/Core/Tool/GaussianMatrix.h
@@ -9,7 +9,7 @@ namespace tezcat::Tiny
 		static float gaussian(float x, float mu, float sigma)
 		{
 			const float a = (x - mu) / sigma;
-			return std::exp(-0.5 * a * a);
+			return (float)std::exp(-0.5f * a * a);
 		}
 
 		static std::vector<float> calculate(int kernelRadius, int& length)
